* fix strange problem with exception ending like this 'ImportError: No module named meade'.
There is a wierd problem in Chimera remote exception handling and the current
way to handle it is to move all your exceptions definition to src/core/exceptions.py module
and import it from your module. I tried to solve this in the past, but was really hard to reproduce.